pyproject.toml
[project]
name = "ragbear"
version = "0.0.1a1"
authors = [{ name = "Oyster Oil", email = "infinitesheldon@gmail.com" }]
description = "Comprehensive RAG implementations"
license = { file = "LICENSE" }
dependencies = [
    "fastapi",
    "torch",
    "loguru",
    "pyyaml",
    "readability-lxml",
    "markdownify",
    "duckduckgo-search",
]
readme = "README.md"
requires-python = ">=3.8"
urls = { "repository" = "https://github.com/oysteroil/ragbear" }
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]


[project.optional-dependencies]
dev = [ ]
build = [ ]
test = [
    "pytest>=8.1.1,<9.0",
]


[build-system]
requires = ["wheel", "setuptools"]
build-backend = "setuptools.build_meta"


[tool.setuptools]
packages = ["ragbear"]
package-data = {}


[tool.pytest.ini_options]
# pythonpath = [ "." ]
asyncio_mode = "auto"


[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = [
    "E",  # pycodestyle
    "F",  # Pyflakes
    "UP",  # pyupgrade
    "B",  # flake8-bugbear
    "SIM",  # flake8-simplify
    "I",  # isort
]

[tool.mypy]
python_version = "3.8"
ignore_missing_imports = true
exclude = ["benchmark/", "docs/", "examples/", "scripts/", "tests/", "build/"]
